worked on book list CRUD
worked on the crud in the list view
currently doesnt look great but working on it
next up is adding author and genre to the new book dialog box
then I'll optimize api calls, probably by adding a multiple ids endpoint
might also use a map so cache duplicate calls

+++ jakarta-primefaces-library-frontend/src/main/java/dto/GenreDto.java
@@ -6,7 +6,7 @@ public class GenreDto {
     private Integer id;
     private String name;
 
-    private Set<Integer> bookIds;
+    private Set<Long> bookIds;
 
     public Integer getId() {
         return id;
@@ -24,11 +24,11 @@ public class GenreDto {
         this.name = name;
     }
 
-    public Set<Integer> getBookIds() {
+    public Set<Long> getBookIds() {
         return bookIds;
     }
 
-    public void setBookIds(Set<Integer> bookIds) {
+    public void setBookIds(Set<Long> bookIds) {
         this.bookIds = bookIds;
     }
